refactor: Moved mergeIngredients to recipe_tools
